pyproject.toml
[tool.poetry]
name = "worldengine-gui"
version = "0.1.0"
description = ""
authors = ["Samuel Bishop <sam@psyx.co>"]

[tool.poetry.dependencies]
python = "^3.7"
PyQt5 = "^5.15.4"
worldengine = {git = "https://github.com/techdragon/worldengine.git"}
numpy = "1.14.6"
PyPlatec = "^1.4.0"
noise = "1.2.2"
protobuf = "3.0.0a3"
pypng = "^0.0.20"
six = "^1.16.0"
setuptools = "^57.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
